ban hang tong dai ly 11:58 30/03/2021

+++ app/src/main/java/vn/gas/thq/ui/retail/RequestInitRetail.kt
@@ -15,11 +15,11 @@ class RequestInitRetail {
 
     @SerializedName("lat")
     @Expose
-    var lat: Int? = null
+    var lat: Float? = null
 
     @SerializedName("lng")
     @Expose
-    var lng: Int? = null
+    var lng: Float? = null
 
     @SerializedName("pay_method")
     @Expose
